Mben进行流量监控

+++ src/main/java/org/netty/mbean/IoAcceptorStatMBean.java
@@ -2,7 +2,17 @@ package org.netty.mbean;
 
 public interface IoAcceptorStatMBean {
 
-	public double getWrittenBytesThroughput();
-	
-	public double getReadBytesThroughput();
+	/**
+	 * netty写流量 bytes/s 对浏览器来说其实就是下载速度
+	 * 
+	 * @return
+	 */
+	public long getWrittenBytesThroughput();
+
+	/**
+	 * netty 读流量 bytes/s 对浏览器来说其实就是上传速度
+	 * 
+	 * @return
+	 */
+	public long getReadBytesThroughput();
 }
